修改了传入课程序号格式要求，限制删除依赖项，增加了修改课程申请时的状态检查

+++ src/main/java/com/example/demo/service/serviceImpl/UserServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.demo.service.serviceImpl;
 import com.example.demo.bean.User;
 import com.example.demo.mapper.UserMapper;
 import com.example.demo.service.GeneralService;
+import com.example.demo.utils.ConstraintsVerify;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -74,6 +75,8 @@ public class UserServiceImpl implements GeneralService<User> {
         User user = userMapper.findByUserid(userid);
 
         if (user != null) {
+            if (ConstraintsVerify.UserHavingDependency(user))
+                return "DependError";
             userMapper.delete(user);
             return "Success";
         } else {
